Initial (buggy) implementation of the infinite part of infinite universe
there appears to be some issues with calculating the correct position
for a superstructure region.

optimization is going to be a PITA.
(did improve the camera system to be more automated setup, so I can experiment with different camera depth / camera counts)

one area of optimization that could be a big help is to keep the number of UniverseEntities at a minimum at all times.
the current behavior of IntergalacticSuperstructure is to immediatly expand into filaments.  this can be optimized by instead calculating nodes/segments
and having them be child objects until the observer is close enough, and only then trigger the expand, and it is when expanding that all the segments and nodes
are converted into individual UniverseEntities.